+++ app/admin/admin_user.rb
@@ -10,14 +10,15 @@ ActiveAdmin.register AdminUser do
     controller.current_ability.can?(:manage, AdminUser)
   }
   form partial: "form"
+
   show title: :title do
     attributes_table do
       row :id
       row :email
       row :name
       row :registration_role
-      row :provider
       row :time_zone
+      row :provider
       row 'Facebook Page', &:facebook
       row :twitter
       row :diaspora
@@ -31,27 +32,46 @@ ActiveAdmin.register AdminUser do
       row "Logged in at", &:last_sign_in_at
       row "Last Seen", &:current_sign_in_at
       row "Last IP" do |user|
-        link_to user.last_sign_in_ip,
-          "http://www.geoiptool.com/en/?IP=#{user.last_sign_in_ip}"
+        if user.last_sign_in_ip
+          link_to user.last_sign_in_ip.to_s, 
+            "http://www.geoiptool.com/en/?IP=#{user.last_sign_in_ip.to_s}"
+        end
       end
       row "Current IP" do |user|
-        link_to user.current_sign_in_ip,
-          "http://www.geoiptool.com/en/?IP=#{user.current_sign_in_ip}"
+        if user.current_sign_in_ip
+          link_to user.current_sign_in_ip.to_s,
+            "http://www.geoiptool.com/en/?IP=#{user.current_sign_in_ip.to_s}"
+        end
       end
       row :created_at
       row :updated_at
+      row "Voted Karma" do |user|
+        user.karma
+      end
+      row "Comments Karma" do |user|
+        user.comments_karma
+      end
+      row "Total Site Karma" do |user|
+        user.full_karma
+      end
+      row "Trusted for comments" do |user|
+        bool_symbol user.comments_trusted?
+      end
     end
     render "user_comments"
   end
+
   index title: "Admin Users" do
     id_column
     column "Avatar", sortable: false do |user|
-      image_tag(user.main_image(:thumb))
+      image_tag(user.main_image(:thumb), class: 'user-avatar-mini')
     end
     column :name
-    column :email
+    column :email do |user|
+      user.email.truncate(20)
+    end
+    column :karma, sortable: false
     column :provider
-    column :time_zone
     column "Facebook", sortable: false do |user|
       if user.facebook.present?
         link_to "Facebook", user.facebook
@@ -71,12 +91,16 @@ ActiveAdmin.register AdminUser do
     column "Logged in", :last_sign_in_at
     column "Last Seen", :current_sign_in_at
     column "Last IP" do |user|
-      link_to user.last_sign_in_ip,
-        "http://www.geoiptool.com/en/?IP=#{user.last_sign_in_ip}"
+      if user.last_sign_in_ip
+        link_to user.last_sign_in_ip.to_s,
+          "http://www.geoiptool.com/en/?IP=#{user.last_sign_in_ip.to_s}"
+      end
     end
     column "Current IP" do |user|
-      link_to user.current_sign_in_ip,
-        "http://www.geoiptool.com/en/?IP=#{user.current_sign_in_ip}"
+      if user.current_sign_in_ip
+        link_to user.current_sign_in_ip.to_s,
+          "http://www.geoiptool.com/en/?IP=#{user.current_sign_in_ip.to_s}"
+      end
     end
     default_actions
   end
@@ -86,7 +110,7 @@ ActiveAdmin.register AdminUser do
       authorize! :update, @admin_user
       respond_to do |format|
         if @admin_user.update_without_password(params[:admin_user])
-          format.html { redirect_to admin_users_path, notice: 'AdminItem was successfully updated.' }
+          format.html { redirect_to admin_users_path, notice: 'User was successfully updated.' }
           format.json { head :ok }
         else
           format.html { render action: "edit" }
@@ -96,6 +120,9 @@ ActiveAdmin.register AdminUser do
     end
     def scoped_collection
       AdminUser.includes(:subscriptions, :roles)
+    end
+    rescue_from CanCan::AccessDenied do |exception|
+      redirect_to admin_access_denied_path, alert: exception.message
     end
   end
 end