+++ app/admin/item_stat.rb
@@ -24,5 +24,8 @@ ActiveAdmin.register ItemStat do
     def scoped_collection
        ItemStat.includes(:item)
     end
+    rescue_from CanCan::AccessDenied do |exception|
+      redirect_to admin_access_denied_path, alert: exception.message
+    end
   end
 end