+++ app/assets/javascripts/active_admin.js
@@ -13,3 +13,4 @@
 //= require easyTooltip
 //= require tooltip
 //= require tinymce/plugins/uploadimage/editor_plugin.js
+//= require highcharts