+++ app/helpers/users_helper.rb
@@ -4,4 +4,6 @@ module UsersHelper
     user.main_image(:thumb)
   end
 
+  def karma_medals(user)
+  end
 end