+++ db/migrate/20121208012116_add_protected_to_items.rb
@@ -1,0 +1,5 @@
+class AddProtectedToItems < ActiveRecord::Migration
+  def change
+    add_column :items, :protected, :boolean, default: false
+  end
+end